Rename Legacy Launcher Api classes
