+++ CollapseLauncher/Classes/Helper/LauncherApiLoader/ILauncherApi.cs
@@ -1,6 +1,6 @@
 ﻿using CollapseLauncher.Extension;
 using CollapseLauncher.Helper.LauncherApiLoader.HoYoPlay;
-using CollapseLauncher.Helper.LauncherApiLoader.Sophon;
+using CollapseLauncher.Helper.LauncherApiLoader.Legacy;
 using System;
 using System.Net.Http;
 using System.Threading;
